Update with tests
